pyproject.toml
[build-system]
requires = ['setuptools >= 80.9.0']
build-backend = 'setuptools.build_meta'

[tool.setuptools]
packages = ['diffres']

[project]
name = 'diffres'
version = '0.0.1'
authors = [
    { name='Zheng Zhao', email='zz@zabemon.com' },
]
description = 'Differentialble resampling with ensemble score'
readme = 'README.md'
license = {file = 'LICENSE'}
requires-python = '>=3.10'
classifiers = [
    'Programming Language :: Python :: 3',
    'License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)',
    'Operating System :: OS Independent',
    'Intended Audience :: Science/Research',
    'Intended Audience :: Developers',
    'Topic :: Scientific/Engineering :: Artificial Intelligence',
    'Topic :: Scientific/Engineering :: Information Analysis',
    'Topic :: Scientific/Engineering :: Mathematics'
]
keywords = ['Markov models']
dependencies = [
    'pytest',
    'numpy==2.3.3',
    'jax==0.7.2',
    'jaxlib==0.7.2',
    'scipy==1.16.2',
    'matplotlib==3.10.6',
    'ott-jax==0.5.2',
    'matplotlib'
]

[project.urls]
homepage = 'https://github.com/zgbkdlm/diffres'
documentation = 'https://github.com/zgbkdlm/diffres'
repository = 'https://github.com/zgbkdlm/diffres'
